last commit before i leave school 21

+++ src/model/app/includes/metrics.h
@@ -5,10 +5,10 @@
 
 namespace s21 {
 typedef struct metrics {
-  float average_accuracy;
-  float average_precision;
-  float average_recall;
-  float average_f_measure;
+  double average_accuracy;
+  double average_precision;
+  double average_recall;
+  double average_f_measure;
   time_t time_elapsed;
 } metrics;
 }  // namespace s21
