+++ src/model/picture/picture_scaler.cc
@@ -18,16 +18,16 @@ Picture PictureScaler::ScalePicture(Picture const &picture, int new_width,
       result_pic[y] = vector<Pixel>(new_width);
 
       for (int x = 0; x < new_width; x++) {
-        float source_x = (x * old_width) / static_cast<float>(new_width);
-        float source_y = (y * old_height) / static_cast<float>(new_height);
+        double source_x = (x * old_width) / static_cast<double>(new_width);
+        double source_y = (y * old_height) / static_cast<double>(new_height);
 
         int x0 = static_cast<int>(source_x);
         int x1 = std::min(x0 + 1, old_width - 1);
         int y0 = static_cast<int>(source_y);
         int y1 = std::min(y0 + 1, old_height - 1);
 
-        float tx = source_x - x0;
-        float ty = source_y - y0;
+        double tx = source_x - x0;
+        double ty = source_y - y0;
 
         int interpolatedValue = (1 - tx) * (1 - ty) * picture(y0, x0) +
                                 tx * (1 - ty) * picture(y0, x1) +
